// src/lib/api/generated/model/depositMode.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Turnerito API
 * API del turnero SaaS multi-tenant
 * OpenAPI spec version: 1.0
 */

export type DepositMode = typeof DepositMode[keyof typeof DepositMode];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DepositMode = {
  none: 'none',
  required: 'required',
  hybrid: 'hybrid',
} as const;